ComandController et comand.html

+++ src/main/java/fr/eni/PizzaOnlinenico/dal/PizzaDAO.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import fr.eni.PizzaOnlinenico.bo.Pizza;
 
-public interface PizzaDAO extends CrudRepository<Pizza, Integer>{
+public interface PizzaDAO extends CrudRepository<Pizza, Long>{
